Give ShuttleGui helper its own file
